initial implementation of COCO loader
